refactor created DateStringFormatter class to test
