+++ app/src/main/java/com/example/insightsshare/EventplanningActivity5.java
@@ -46,6 +46,8 @@ public class EventplanningActivity5 extends AppCompatActivity {
     private Button timePickerButton;
     int tHour, tMin;
 
+    private DateStringFormatter dateStringFormatter = new DateStringFormatter();
+
     //variables for transferring Eventdata to DB
     TextView eventCreator;
     EditText eventName, eventDescription, eventPlace, maxParticipants;
@@ -228,7 +230,7 @@ public class EventplanningActivity5 extends AppCompatActivity {
         int month = cal.get(Calendar.MONTH);
         month=month+1;
         int day = cal.get(Calendar.DAY_OF_MONTH);
-        return makeDateString(day, month, year);
+        return dateStringFormatter.makeDateString(day, month, year);
     }
 
 
@@ -239,7 +241,7 @@ public class EventplanningActivity5 extends AppCompatActivity {
 
             //so Jan is 1 instead of 0
             month=month+1;
-            String date= makeDateString(day, month, year);
+            String date= dateStringFormatter.makeDateString(day, month, year);
             datePickerButton.setText(date);
         };
 
@@ -292,46 +294,6 @@ public class EventplanningActivity5 extends AppCompatActivity {
         //Display previously selected time
         timePickerDialog.updateTime(tHour, tMin);
     } //end of initTimePicker
-
-
-    //for DatePicker: called in methode initDatePicker()
-    private String makeDateString(int day, int month, int year) {
-
-        return getMonthFormat(month) + " " + day +  " " +year;
-    }
-
-
-    //for DatePicker: called in methode makeDateString()
-    private String getMonthFormat(int month) {
-
-        if(month==1)
-            return "JAN";
-        if(month==2)
-            return "FEB";
-        if(month==3)
-            return "MAR";
-        if(month==4)
-            return "APR";
-        if(month==5)
-            return "MAY";
-        if(month==6)
-            return "JUN";
-        if(month==7)
-            return "AUG";
-        if(month==8)
-            return "SEP";
-        if(month==9)
-            return "SEP";
-        if(month==10)
-            return "OCT";
-        if(month==11)
-            return "NOV";
-        if(month==12)
-            return "DEZ";
-
-        //default should never happen
-        return "JAN";
-    }
 
     public void openDatePicker(View view){
         datePickerDialog.show();
